% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DICOM_codes.R
\docType{data}
\name{DICOM_codes}
\alias{DICOM_codes}
\title{rda containing DICOM header codes to include in \emph{RIA_image} object}
\format{Each row is a DICOM header input}
\usage{
DICOM_codes
}
\value{
3 column data.frame
}
\description{
rda data file containing Name, Group and Element codes of DICOM header info
to be included into \emph{RIA_image} object by default when using \code{\link[RIA]{load_dicom}} function.
Can be edited to change defaults.
}

