% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Non_NRS.R
\docType{data}
\name{Non_NRS}
\alias{Non_NRS}
\title{\emph{RIA_image} object of a plaque without the napkin-ring sign}
\format{RIA_image object}
\usage{
NRS
}
\value{
RIA_image object
}
\description{
rda containing an example \emph{RIA_image} object of a patients plaque
which does not show the napkin-ring sign.
}

