% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NRS.R
\docType{data}
\name{NRS}
\alias{NRS}
\title{\emph{RIA_image} object of a plaque with the napkin-ring sign}
\format{RIA_image object}
\usage{
NRS
}
\value{
RIA_image object
}
\description{
rda containing an example \emph{RIA_image} object of a patients plaque
which shows the napkin-ring sign.
}

