% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DICOM_codes.R
\docType{data}
\encoding{UTF-8}
\name{DICOM_codes}
\alias{DICOM_codes}
\title{rda containing DICOM header codes to include in \emph{RIA_image} object}
\format{Each row is a DICOM header input}
\usage{
DICOM_codes
}
\value{
3 column data.frame
}
\description{
rda data file containing Name, Group and Element codes of DICOM header info
to be included into \emph{RIA_image} object by default when using \code{\link[RIA]{load_dicom}} function.
Can be edited to change defaults.
}
\references{
Márton KOLOSSVÁRY et al.
Radiomic Features Are Superior to Conventional Quantitative Computed Tomographic
Metrics to Identify Coronary Plaques With Napkin-Ring Sign
Circulation: Cardiovascular Imaging (2017).
DOI: 10.1161/circimaging.117.006843
\url{http://circimaging.ahajournals.org/content/10/12/e006843}

Márton KOLOSSVÁRY et al.
Cardiac Computed Tomography Radiomics: A Comprehensive Review on Radiomic Techniques.
Journal of Thoracic Imaging (2017).
DOI: 10.1097/RTI.0000000000000268
\url{https://www.ncbi.nlm.nih.gov/pubmed/28346329}
}
