% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glrlm_stat.R
\encoding{UTF-8}
\name{glrlm_stat}
\alias{glrlm_stat}
\title{GLRLM-based statistics}
\usage{
glrlm_stat(
  RIA_data_in,
  use_type = "single",
  use_orig = FALSE,
  use_slot = "glrlm$es_8_111",
  save_name = NULL,
  verbose_in = TRUE
)
}
\arguments{
\item{RIA_data_in}{\emph{RIA_image}.}

\item{use_type}{string, can be \emph{"single"} which runs the function on a single image,
which is determined using \emph{"use_orig"} or \emph{"use_slot"}. \emph{"glrlm"}
takes all datasets in the \emph{RIA_image$glrlm} slot and runs the analysis on them.}

\item{use_orig}{logical, indicating to use image present in \emph{RIA_data$orig}.
If FALSE, the modified image will be used stored in \emph{RIA_data$modif}. However, GLRLM matrices
are usually note present in either slots, therefore giving the slot name using \emph{use_slot} is advised.}

\item{use_slot}{string, name of slot where data wished to be used is. Use if the desired image
is not in the \emph{data$orig} or \emph{data$modif} slot of the \emph{RIA_image}. For example,
ig the desired dataset is in \emph{RIA_image$glrlm$ep_4}, then \emph{use_slot} should be
\emph{glrlm$ep_4}. The results are automatically saved. If the results are not saved to
the desired slot, then please use \emph{save_name} parameter.}

\item{save_name}{string, indicating the name of subslot of \emph{$glrlm} to save results to.
If left empty, then it will be automatically determined.}

\item{verbose_in}{logical, indicating whether to print detailed information.
Most prints can also be suppressed using the \code{\link{suppressMessages}} function.}
}
\value{
\emph{RIA_image} containing the statistical information.
}
\description{
Calculates GLRLM-based statistics for given GLRLM matrix.
}
\examples{
\dontrun{
#Discretize loaded image and then calculate GLRLM statistics
RIA_image <- discretize(RIA_image, bins_in = 8, equal_prob = TRUE)
RIA_image <- glrlm(RIA_image, use_orig = FALSE, use_slot = "discretized$ep_8",
right = TRUE, down = TRUE, forward = FALSE)
RIA_image <- glrlm_stat(RIA_image, use_orig = FALSE, use_slot = "glrlm$ep_8_110")

#Batch calculation of GLRLM-based statistics on all calculated GLRLMs
RIA_image <- glrlm_stat(RIA_image, use_type = "discretized")
}

}
\references{
Márton KOLOSSVÁRY et al.
Radiomic Features Are Superior to Conventional Quantitative Computed Tomographic
Metrics to Identify Coronary Plaques With Napkin-Ring Sign
Circulation: Cardiovascular Imaging (2017).
DOI: 10.1161/circimaging.117.006843
\url{https://pubmed.ncbi.nlm.nih.gov/29233836/}

Márton KOLOSSVÁRY et al.
Cardiac Computed Tomography Radiomics: A Comprehensive Review on Radiomic Techniques.
Journal of Thoracic Imaging (2018).
DOI: 10.1097/RTI.0000000000000268
\url{https://pubmed.ncbi.nlm.nih.gov/28346329/}
}
