% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Non_NRS.R
\docType{data}
\encoding{UTF-8}
\name{Non_NRS}
\alias{Non_NRS}
\title{\emph{RIA_image} object of a plaque without the napkin-ring sign}
\format{
RIA_image object
}
\usage{
NRS
}
\value{
RIA_image object
}
\description{
rda containing an example \emph{RIA_image} object of a patients plaque
which does not show the napkin-ring sign.
}
\references{
Márton KOLOSSVÁRY et al.
Radiomic Features Are Superior to Conventional Quantitative Computed Tomographic
Metrics to Identify Coronary Plaques With Napkin-Ring Sign
Circulation: Cardiovascular Imaging (2017).
DOI: 10.1161/circimaging.117.006843
\url{https://pubmed.ncbi.nlm.nih.gov/29233836/}

Márton KOLOSSVÁRY et al.
Cardiac Computed Tomography Radiomics: A Comprehensive Review on Radiomic Techniques.
Journal of Thoracic Imaging (2018).
DOI: 10.1097/RTI.0000000000000268
\url{https://pubmed.ncbi.nlm.nih.gov/28346329/}
}
